Refactored MessageSystem to actually be true messages, instead of a weird message/command hybrid. Now there are commands which are used in ViewModels by their Views.
Commands change the state of the ViewModel or Model. Views will update accordingly, by way of the INotifyPropertyChanged event.

Messages are used between ViewModels to signal events that might affect more than one View. For example, when a player hits the scroll trigger, the environment, player, enemies, and powerups should all move up. This might be handled when a ScrollMessage is received.